+++ src/beshell/JSEngine.hpp
@@ -84,6 +84,8 @@ namespace be {
         static void setGlobalValue(JSContext * ctx, const char * name, JSValue) ;
         static JSValue getGlobalValue(const char * name) ;
 
+        static constexpr int FLAG_EXPORT_TO_GLOBAL = 0x100 ;
+
         JSValue eval(const char * code, int code_len=-1, const char * filename="eval", int flags=JS_EVAL_TYPE_GLOBAL) ;
         JSValue evalScript(const char * filepath, int flags=JS_EVAL_TYPE_MODULE, bool dumpException=true) ;